KIALI-1985 Refactor IstioObjects to map metadata/spec properly
