+++ src/pages/ServiceDetails/ServiceInfo/__tests__/ServiceInfoRouteVirtualServices.test.tsx
@@ -1,56 +1,58 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import { VirtualService } from '../../../../types/ServiceInfo';
 import ServiceInfoVirtualServices from '../ServiceInfoVirtualServices';
+import { VirtualService } from '../../../../types/IstioObjects';
 
 const virtualServices: VirtualService[] = [
   {
-    name: 'reviews-default',
-    createdAt: '2018-03-14T10:17:52Z',
-    resourceVersion: '1234',
-    hosts: ['rewiews'],
-    gateways: ['reviews'],
-    http: [
-      {
-        route: [
-          {
-            destination: {
-              subset: 'v1',
-              host: 'reviews'
+    metadata: {
+      name: 'reviews-default'
+    },
+    spec: {
+      hosts: ['rewiews'],
+      gateways: ['reviews'],
+      http: [
+        {
+          route: [
+            {
+              destination: {
+                subset: 'v1',
+                host: 'reviews'
+              },
+              weight: 55
             },
-            weight: 55
-          },
-          {
-            destination: {
-              subset: 'v3',
-              host: 'reviews'
+            {
+              destination: {
+                subset: 'v3',
+                host: 'reviews'
+              },
+              weight: 55
+            }
+          ]
+        }
+      ],
+      tcp: [
+        {
+          match: [],
+          route: [
+            {
+              destination: {
+                subset: 'v1',
+                host: 'reviews'
+              },
+              weight: 55
             },
-            weight: 55
-          }
-        ]
-      }
-    ],
-    tcp: [
-      {
-        match: [],
-        route: [
-          {
-            destination: {
-              subset: 'v1',
-              host: 'reviews'
-            },
-            weight: 55
-          },
-          {
-            destination: {
-              subset: 'v2',
-              host: 'reviews'
-            },
-            weight: 55
-          }
-        ]
-      }
-    ]
+            {
+              destination: {
+                subset: 'v2',
+                host: 'reviews'
+              },
+              weight: 55
+            }
+          ]
+        }
+      ]
+    }
   }
 ];
