+++ src/types/ServiceInfo.ts
@@ -1,7 +1,6 @@
 import { ServiceHealth } from './Health';
 import { PfColors } from '../components/Pf/PfColors';
-import { Pod, Port, ObjectCheck, ObjectValidation } from './IstioObjects';
-import { ResourcePermissions } from './Permissions';
+import { Pod, Port, ObjectCheck, ObjectValidation, DestinationRules, VirtualServices } from './IstioObjects';
 
 export interface Endpoints {
   addresses?: EndpointAddress[];
@@ -20,327 +19,6 @@ export interface WorkloadOverview {
   labels?: { [key: string]: string };
   resourceVersion: string;
   createdAt: string;
-}
-
-// RouteRule type
-
-export interface RouteRule {
-  name: string;
-  createdAt: string;
-  resourceVersion: string;
-  destination?: IstioService;
-  precedence?: number;
-  match?: MatchCondition;
-  route?: DestinationWeightV1Alpha1[];
-  redirect?: HTTPRedirect;
-  rewrite?: HTTPRewrite;
-  websocketUpgrade?: string;
-  httpReqTimeout?: HTTPTimeout;
-  httpReqRetries?: HTTPRetry;
-  httpFault?: HTTPFaultInjection;
-  l4Fault?: L4FaultInjection;
-  mirror?: IstioService;
-  corsPolicy?: CorsPolicy;
-  appendHeaders?: { [key: string]: string };
-}
-
-export interface IstioService {
-  name?: string;
-  namespace?: string;
-  domain?: string;
-  service?: string;
-  labels?: { [key: string]: string };
-}
-
-export interface MatchCondition {
-  source?: IstioService;
-  tcp?: L4MatchAttributes;
-  udp?: L4MatchAttributes;
-  request?: MatchRequest;
-}
-
-export interface L4MatchAttributes {
-  sourceSubnet: string[];
-  destinationSubnet: string[];
-}
-
-export interface TLSMatchAttributes {
-  sniHosts: string[];
-  destinationSubnet: string[];
-  port: number;
-  sourceLabels: { [key: string]: string };
-  gateways: string[];
-}
-
-export interface MatchRequest {
-  headers: { [key: string]: StringMatch };
-}
-
-export interface StringMatch {
-  exact?: string;
-  prefix?: string;
-  regex?: string;
-}
-
-export interface DestinationWeightV1Alpha1 {
-  labels: { [key: string]: string };
-  weight?: number;
-}
-
-export interface DestinationWeight {
-  destination: Destination;
-  weight?: number;
-}
-
-export interface HTTPRedirect {
-  uri: string;
-  authority: string;
-}
-
-export interface HTTPRewrite {
-  uri: string;
-  authority: string;
-}
-
-export interface HTTPTimeout {
-  simpleTimeout: SimpleTimeoutPolicy;
-  custom: string;
-}
-
-export interface SimpleTimeoutPolicy {
-  timeout: string;
-  overrideHeaderName: string;
-}
-
-export interface HTTPRetry {
-  simpleRetry: SimpleRetryPolicy;
-  custom: string;
-}
-
-export interface SimpleRetryPolicy {
-  attempts: number;
-  perTryTimeout: string;
-  overrideHeaderName: string;
-}
-
-export interface HTTPFaultInjection {
-  delay: Delay;
-  abort: Abort;
-}
-
-export interface Delay {
-  percent: number;
-  fixedDelay: string;
-  exponentialDelay: string;
-  overrideHeaderName: string;
-}
-
-export interface Abort {
-  percent: number;
-  grpcStatus: string;
-  http2Error: string;
-  httpStatus: string;
-  overrideHeaderName: string;
-}
-
-export interface L4FaultInjection {
-  throttle: Throttle;
-  terminate: Terminate;
-}
-
-export interface Throttle {
-  percent: number;
-  downstreamLimitBps: number;
-  upstreamLimitBps: number;
-  throttleAfterPeriod: string;
-  throttleAfterBytes: number;
-  throttleForPeriod: string;
-}
-
-export interface Terminate {
-  percent: number;
-  terminateAfterPeriod: string;
-}
-
-export interface CorsPolicy {
-  allowOrigin: string[];
-  allowMethods: string[];
-  allowHeaders: string[];
-  exposeHeaders: string[];
-  maxAge: string;
-  allowCredentials: string;
-}
-
-// Destination Policy
-
-export interface LoadBalancing {
-  name: string;
-}
-
-export interface CircuitBreakerPolicy {
-  maxConnections?: number;
-  httpMaxPendingRequests?: number;
-  httpMaxRequests?: number;
-  sleepWindow?: string;
-  httpConsecutiveErrors?: string;
-  httpDetectionInterval?: string;
-  httpMaxRequestsPerConnection?: number;
-  httpMaxEjectionPercent?: number;
-  httpMaxRetries?: number;
-}
-
-export interface CircuitBreaker {
-  simpleCb: CircuitBreakerPolicy;
-  custom: string;
-}
-
-export interface DestinationPolicy {
-  name: string;
-  createdAt: string;
-  resourceVersion: string;
-  destination?: IstioService;
-  source?: IstioService;
-  loadbalancing?: LoadBalancing;
-  circuitBreaker?: CircuitBreaker;
-}
-
-// Virtual Service
-
-export interface PortSelector {
-  number: number;
-  name: string;
-}
-
-export interface Destination {
-  host: string;
-  subset?: string;
-  port?: PortSelector;
-}
-
-export interface HTTPMatchRequest {
-  uri: StringMatch;
-  scheme: StringMatch;
-  method: StringMatch;
-  authority: StringMatch;
-  headers: { [key: string]: StringMatch };
-  port: PortSelector;
-  sourceLabels: { [key: string]: string };
-  gateways: string[];
-}
-
-export interface HTTPRoute {
-  match?: HTTPMatchRequest[];
-  route?: DestinationWeight[];
-  redirect?: HTTPRedirect;
-  rewrite?: HTTPRewrite;
-  websocketUpgrade?: boolean;
-  timeout?: string;
-  retries?: HTTPRetry;
-  mirror?: Destination;
-  corsPolicy?: CorsPolicy;
-  appendHeaders?: { [key: string]: string };
-}
-
-export interface TCPRoute {
-  match?: L4MatchAttributes[];
-  route?: DestinationWeight[];
-}
-
-export interface TLSRoute {
-  match?: TLSMatchAttributes[];
-  route?: DestinationWeight[];
-}
-
-export interface VirtualService {
-  name: string;
-  createdAt: string;
-  resourceVersion: string;
-  hosts?: string[];
-  gateways?: string[];
-  http?: HTTPRoute[];
-  tcp?: TCPRoute[];
-  tls?: TLSRoute[];
-}
-
-export interface VirtualServices {
-  items: VirtualService[];
-  permissions: ResourcePermissions;
-}
-
-// Destination Rule
-
-export interface ConsistentHashLB {
-  httpHeader: string;
-  minimumRingSize: number;
-}
-
-export interface LoadBalancerSettings {
-  simple: string;
-  consistentHash: ConsistentHashLB;
-}
-
-export interface ConnectionPoolSettingsTCPSettings {
-  maxConnections: number;
-  connectTimeout: string;
-}
-
-export interface ConnectionPoolSettingsHTTPSettings {
-  http1MaxPendingRequests: number;
-  http2MaxRequests: number;
-  maxRequestsPerConnection: number;
-  maxRetries: number;
-}
-
-export interface ConnectionPoolSettings {
-  tcp: ConnectionPoolSettingsTCPSettings;
-  http: ConnectionPoolSettingsHTTPSettings;
-}
-
-export interface OutlierDetectionHTTPSettings {
-  consecutiveErrors: number;
-  interval: string;
-  baseEjectionTime: string;
-  maxEjectionPercent: number;
-}
-
-export interface OutlierDetection {
-  http: OutlierDetectionHTTPSettings;
-}
-
-export interface TLSSettings {
-  mode: string;
-  clientCertificate: string;
-  privateKey: string;
-  caCertificates: string;
-  subjectAltNames: string[];
-  sni: string;
-}
-
-export interface TrafficPolicy {
-  loadBalancer: LoadBalancerSettings;
-  connectionPool: ConnectionPoolSettings;
-  outlierDetection: OutlierDetection;
-  tls: TLSSettings;
-}
-
-export interface Subset {
-  name: string;
-  labels: { [key: string]: string };
-  trafficPolicy: TrafficPolicy;
-}
-
-export interface DestinationRule {
-  name: string;
-  createdAt: string;
-  resourceVersion: string;
-  host?: string;
-  trafficPolicy?: TrafficPolicy;
-  subsets?: Subset[];
-}
-
-export interface DestinationRules {
-  items: DestinationRule[];
-  permissions: ResourcePermissions;
 }
 
 export interface SourceWorkload {
